+++ FastAnimations/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## 1.6
 * Updated for Stardew Valley 1.3.
+* Added support for title menu transitions.
 
 ## 1.5
 * Updated to SMAPI 2.4.